requirements.txt
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
Django==5.1.3
django-cors-headers==4.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
h11==0.14.0
idna==3.10
inflection==0.5.1
outcome==1.3.0.post0
packaging==24.1
Pillow==10.0.1
psycopg2-binary==2.9.7
pycparser==2.22
PyJWT==2.9.0
PySocks==1.7.1
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.31.0
selenium==4.26.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.1
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
